refactor: extract polaroid component
