+++ src/app/pages/Designathons/Designathon24/components/Polaroid/Polaroid.module.scss
@@ -1,0 +1,75 @@
+.polaroid {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	position: relative;
+
+	width: 100%;
+
+	padding: 0.35rem 0.35rem;
+
+	.polaroidImage {
+		width: 100%;
+		height: 100%;
+
+		border-radius: 0.25rem;
+	}
+
+	.polaroidDetails {
+		padding-top: 0.5rem;
+		text-align: left;
+
+		.polaroidName {
+			font-size: 1rem;
+			font-weight: 600;
+
+			padding-bottom: 8px;
+		}
+
+		.polaroidPosition {
+			font-size: 0.75rem;
+			font-weight: 400;
+		}
+	}
+
+	.tape {
+		width: 66%;
+		z-index: 5;
+
+		position: absolute;
+		top: -5%;
+		left: 50%;
+		transform: translateX(-50%);
+	}
+
+	&:hover {
+		background-color: #fffbf3;
+		box-shadow: 0px 4px 6px 0 #00000030;
+		transition: 0.1s ease-in;
+	}
+
+	@media screen and (min-width: 640px) {
+		.polaroidDetails {
+			.polaroidName {
+				font-size: 2rem;
+			}
+
+			.polaroidPosition {
+				font-size: 1.25rem;
+			}
+		}
+	}
+
+	@media screen and (min-width: 768px) {
+		width: 300px;
+		padding: 0.75rem 0.75rem;
+	}
+
+	@media screen and (min-width: 1024px) {
+		width: 400px;
+	}
+
+	@media screen and (min-width: 1728px) {
+		width: 450px;
+	}
+}
