+++ src/app/pages/Designathons/Designathon24/components/Speakers/Speakers.module.scss
@@ -31,57 +31,6 @@
 			padding-bottom: 3rem;
 		}
 
-		.polaroid {
-			display: flex;
-			flex-direction: column;
-			justify-content: center;
-			position: relative;
-
-			width: 100%;
-
-			padding: 0.35rem 0.35rem;
-
-			.polaroidImage {
-				width: 100%;
-				height: 100%;
-
-				border-radius: 0.25rem;
-			}
-
-			.polaroidDetails {
-				padding-top: 0.5rem;
-				text-align: left;
-
-				.polaroidName {
-					font-size: 1rem;
-					font-weight: 600;
-
-					padding-bottom: 8px;
-				}
-
-				.polaroidPosition {
-					font-size: 0.75rem;
-					font-weight: 400;
-				}
-			}
-
-			.tape {
-				width: 66%;
-				z-index: 5;
-
-				position: absolute;
-				top: -5%;
-				left: 50%;
-				transform: translateX(-50%);
-			}
-
-			&:hover {
-				background-color: #fffbf3;
-				box-shadow: 0px 4px 6px 0 #00000030;
-				transition: 0.1s ease-in;
-			}
-		}
-
 		.keynote {
 			display: flex;
 			flex-direction: column;
@@ -141,20 +90,6 @@
 
 			padding: 0 3rem;
 		}
-
-		.polaroids {
-			.polaroid {
-				.polaroidDetails {
-					.polaroidName {
-						font-size: 2rem;
-					}
-
-					.polaroidPosition {
-						font-size: 1.25rem;
-					}
-				}
-			}
-		}
 	}
 
 	@media screen and (min-width: 768px) {
@@ -163,12 +98,6 @@
 		}
 
 		.polaroids {
-			.polaroid {
-				width: 300px;
-
-				padding: 0.75rem 0.75rem;
-			}
-
 			.peopleContainer {
 				margin: auto;
 
@@ -190,10 +119,6 @@
 			.polaroidHeading {
 				font-size: 2.5rem;
 				padding-bottom: 5rem;
-			}
-
-			.polaroid {
-				width: 400px;
 			}
 
 			.peopleContainer {
@@ -218,16 +143,6 @@
 	}
 
 	@media screen and (min-width: 1728px) {
-		.polaroids {
-			.polaroid {
-				width: 450px;
-
-				.tape {
-					width: 284px;
-				}
-			}
-		}
-
 		.peopleContainer {
 			width: 1568px;
 		}
